default_translation: use puppet yamlresource rather than resource
The `puppet resource` command is quite limited and cannot cope with
array values and other non-trivial input. So I went ahead and built
a face called `puppet yamlresource` that does the same thing as
`puppet resource`, but reads input in YAML form.

We pass in JSON, as that is still valid YAML. Also be careful to disable
colors in the `ifcmd`, as otherwise the "start of the line" regex will
not match.