dsl: allow translators to explicitly ignore attributes, warn otherwise
It's useful to display meaningful messages for certain attribute values.
For example, file/purge=false does not matter, and should not lead to
a warning. purge=true on the other hand will not quite work in `mgmt`,
so an error should be printed during translation.

Other attributes can be ignored altogether.

Attributes that are in the catalog, but are not mentioned by the
translator code at all, now lead to a warning message as well.