add the pippet translator and make it the new default catch-all
We have a new hack to feed resources back to puppet, using the "pippet"
pseudo resource type in mgmt.

Instead of `exec puppet yamlresource`, the resource data will be passed
through nodes of the pippet type.

This is the default behavior from now on, with the old catch-all only
available through the --no-pippet flag.

This is a breaking change.