catalog_translation: use the Puppet::Configurer module
It became evident that the old approach with the `puppet catalog` face
no longer works. Puppetserver (at least in version 7 and 8) cannot serve
the catalog request from this face (apparently it is made without any
facts data).

In order to overcome this limitation, we now monkey-patch the
Puppet::Configurer class in order to give us a very dumbed down method
for still properly requesting a catalog from the server. The method only
includes the most crucial code paths lifted from the run_internal method
and its calling context. This is hard to test, unfortunately, but it
seems to work, at least using --code or the implicit rest client mode.

When running with --manifest, this does not work at all, because the
Puppet::Configurer is not build to handle this switch. The `puppet
apply` code does not use the Configurer module, and implements a lot of
code right in the Application::Apply class. This is not helpful. That's
why in this case, we continue to rely on the catalog face.